Add files via upload
This presentation shared was used for the purpose of an assessment towards a theoretical class on Physics of Classical and Quantum Information, 2015-16, on 16 December 2015 at IST Lisbon, under the DP-PMI program funded by FCT Portugal.

In our presentation we have included some non-trivial calculation or proof on a selective topic of fault tolerant quantum computation, following Prof. Peter Shor's algorithm, not just definitions and concepts. :-)

Regarding the presentation, our goal is to convey a clear view of the topic, in a comprehensive way, i.e. referring adequately the motivation and applications, and within that trying to choose something that it would be instructive for us to calculate/prove. This does not have a complete proof/calculation: we have given the key structure/steps, and than chosen one step that is explained in more detail.